fix mkfs #1924

+++ lib/python/Components/Harddisk.py
@@ -437,7 +437,7 @@ class Harddisk:
 			big_o_options.append("sparse_super")
 		elif size > 2048:
 			# Over 2GB: 32 i-nodes per megabyte
-			task.args += ["-T", "largefile", "-N", str(size * 32)]
+			task.args += ["-T", "largefile", "-N", str(int(size * 32))]
 		task.args += ["-F", "-F", "-m0", "-O ^metadata_csum", "-O", ",".join(big_o_options), self.partitionPath("1")]
 
 		task = MountTask(job, self)
